rename wizard to wizars_steps
